Add Test for Stats

+++ stats.go
@@ -3,7 +3,7 @@ package firequeue
 // Stats contains firequeue statistics.
 type Stats struct {
 	QueueLength      int
-	BatchLength      int
+	BatchLength      int64
 	Success          int64
 	RetryCount       int64
 	UnretryableError int64
